refactor: modify components casing name
