--- src/components/LoadingSpinner.tsx
@@ -1,9 +1,0 @@
-import { CgSpinner } from 'react-icons/cg'
-
-const LoadingSpinner = () => {
-  return (
-    <CgSpinner className="animate-spin h-4 w-4 text-primary-600 dark:text-primary-400" />
-  )
-}
-
-export default LoadingSpinner
