--- src/components/LoginProviderButton/index.tsx
@@ -1,3 +1,0 @@
-import LoginProviderButton from './LoginProviderButton'
-
-export default LoginProviderButton
