--- src/components/Statistics/index.tsx
@@ -1,4 +1,0 @@
-export * from './GithubStats'
-export * from './StatisticsCard'
-export * from './TotalEndorsements'
-export * from './TotalViews'
