+++ src/components/image-lightbox.tsx
