+++ src/components/login-provider-button/index.tsx
@@ -1,0 +1,3 @@
+import LoginProviderButton from './login-provider-button'
+
+export default LoginProviderButton
