+++ src/components/mdx/image.tsx
@@ -3,7 +3,7 @@ import type { ImageProps as NextImageProps } from 'next/image'
 import NextImage from 'next/image'
 import { useState } from 'react'
 
-import ImageLightBox from '../ImageLightbox'
+import ImageLightBox from '../image-lightbox'
 
 type BaseImageProps = Omit<NextImageProps, 'width' | 'height' | 'fill'>
 type SizeProps = BaseImageProps & { size?: number }
@@ -17,7 +17,7 @@ type ImageProps = (SizeProps | DimensionProps) & {
   shouldOpenLightBox?: boolean
 }
 
-export const Image = ({
+const Image = ({
   shouldOpenLightBox = true,
   className,
   ...props
@@ -58,3 +58,5 @@ export const Image = ({
     </figure>
   )
 }
+
+export default Image
