+++ src/components/scroll-progress-bar.tsx
@@ -1,6 +1,6 @@
 import { useContext, useEffect, useState } from 'react'
 
-import { ScrollContext } from './ScrollObserver'
+import { ScrollContext } from './scroll-observer'
 
 const ScrollProgressBar = () => {
   const [width, setWidth] = useState(0)
