+++ src/components/statistics/index.tsx
@@ -1,0 +1,4 @@
+export * from './github-stats'
+export * from './statistics-card'
+export * from './total-endorsements'
+export * from './total-views'
