+++ src/pages/blog/[slug].tsx
@@ -8,10 +8,10 @@ import type {
 import { useRouter } from 'next/router'
 import React, { useMemo } from 'react'
 
-import { mdxComponents, MdxContent } from '@/components/Mdx'
-import PageSeo from '@/components/PageSeo'
-import ScrollProgressBar from '@/components/ScrollProgressBar'
-import Spinner from '@/components/Spinner'
+import { mdxComponents, MdxContent } from '@/components/mdx'
+import PageSeo from '@/components/page-seo'
+import ScrollProgressBar from '@/components/scroll-progress-bar'
+import Spinner from '@/components/spinner'
 import { siteMetadata } from '@/data'
 import { useMDXComponent } from '@/hooks'
 import { getPosts } from '@/lib/contentlayer'
