+++ src/pages/blog/index.tsx
@@ -3,8 +3,8 @@ import type { InferGetStaticPropsType } from 'next'
 import { useMemo, useState } from 'react'
 import { RiSearch2Line } from 'react-icons/ri'
 
-import PageHeader from '@/components/PageHeader'
-import PageSeo from '@/components/PageSeo'
+import PageHeader from '@/components/page-header'
+import PageSeo from '@/components/page-seo'
 import PostCard from '@/components/post-card'
 import { filterPosts, getPosts } from '@/lib/contentlayer'
