+++ src/pages/endorsements.tsx
@@ -2,11 +2,11 @@ import type { InferGetStaticPropsType } from 'next'
 import { signIn, signOut, useSession } from 'next-auth/react'
 import { useState } from 'react'
 
-import EndorsementsBadge from '@/components/EndorsementsBadge'
+import EndorsementsBadge from '@/components/endorsements-badge'
 import Link from '@/components/link'
-import PageHeader from '@/components/PageHeader'
-import PageSeo from '@/components/PageSeo'
-import Spinner from '@/components/Spinner'
+import PageHeader from '@/components/page-header'
+import PageSeo from '@/components/page-seo'
+import Spinner from '@/components/spinner'
 import { useSkillCategories } from '@/hooks'
 import getSkillCategories from '@/lib/getSkillCategories'
