+++ src/pages/resume.tsx
@@ -4,8 +4,8 @@ import { RiFileTextLine } from 'react-icons/ri'
 import CallToAction from '@/components/call-to-action'
 import Divider from '@/components/divider'
 import Link from '@/components/link'
-import PageHeader from '@/components/PageHeader'
-import PageSeo from '@/components/PageSeo'
+import PageHeader from '@/components/page-header'
+import PageSeo from '@/components/page-seo'
 import { experiences, siteMetadata } from '@/data'
 import { formatDate } from '@/lib/utils'
